Fixed two crashes in concealment, tweaked reserved slots

+++ EssentialsPlugin/Utility/CubeGrid.cs
@@ -1633,7 +1633,7 @@
 					{
 						MyObjectBuilder_MotorAdvancedStator stator = (MyObjectBuilder_MotorAdvancedStator)cubeBlock.GetObjectBuilderCubeBlock( );
 						IMyEntity connectedEntity;
-						if ( MyAPIGateway.Entities.TryGetEntityById( (long) stator.RotorEntityId, out connectedEntity ) )
+						if ( MyAPIGateway.Entities.TryGetEntityById( stator.RotorEntityId, out connectedEntity ) )
 						{
 							IMyCubeGrid parent = (IMyCubeGrid)connectedEntity.Parent;
 							if ( !checkedGrids.Contains( parent ) )
